Fetch petitions until finding 10 items
In 'order=latest' mode (for the Trending tab in ClassicPress), we throw
out any petitions that have zero votes, because these are most likely
petitions that have been bumped by the forum system without anyone
commenting on them.

This may cause the first page of Discourse results to have less than 10
valid petitions in this mode, so fetch more pages until we get 10 items
that can be returned in the API result.